# internal/handlers/slack.go
package handlers

import (
	"net/http"
)

// SlackReply godoc
// @Summary Reply to ticket via Slack
// @Description Add a reply to a ticket from Slack interface
// @Tags slack
// @Security BearerAuth
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Failure 403 {object} map[string]string
// @Router /slack/reply [post]
func SlackReply(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{"message":"Slack reply endpoint - Coming soon"}`))
}

// SlackWebhook godoc
// @Summary Slack webhook endpoint
// @Description Webhook endpoint for Slack bot interactions
// @Tags slack
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Router /slack/webhook [post]
func SlackWebhook(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	w.Write([]byte(`{"message":"Slack webhook endpoint - Coming soon"}`))
}